support nullable numericString()
